pyproject.toml
[project]
name = "smactorio"
version = "0.1.0"
description = "Autonomous Turnkey Software Factory using Spec-Driven Development and Agentic Orchestration"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "Leon Breukelman" }
]
keywords = ["ai", "software-factory", "spec-driven-development", "dspy", "agentic"]

dependencies = [
    "dspy-ai>=2.5.0",
    "pydantic>=2.0.0",
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["factory"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "E402",  # module level import not at top of file (for scripts)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["factory", "src"]
branch = true

[tool.coverage.report]
fail_under = 85
show_missing = true
